# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the project.",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get all banners",
                "operationId": "9f718c7db0d13e4b5be26b80595f4cf1",
                "responses": {
                    "200": {
                        "description": "List of banners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Banner"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Create a new banner",
                "operationId": "546707dc3442c51e694ec312df3664e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the banner",
                                        "type": "string",
                                        "example": "Banner 1"
                                    },
                                    "image": {
                                        "description": "Image file for the banner",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "Description of the banner",
                                        "type": "string",
                                        "example": "This is a banner description."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Banner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Banner"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get a specific banner by ID",
                "operationId": "13d01a36bda0d8b491a382619c6bc469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Banner"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Banners"
                ],
                "summary": "Update a specific banner by ID",
                "operationId": "89bb1eb70cfa7909040b083ebe79f6df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Banner Name"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "updated_banner.jpg"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Banner updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Banner"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete a specific banner by ID",
                "operationId": "00a27f787ec0124f0b98cfdeecf6090a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner deleted successfully"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Banner": {
                "title": "Banner",
                "description": "Banner model with name, image, and description",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the banner",
                        "type": "string",
                        "example": "Summer Sale Banner"
                    },
                    "image": {
                        "title": "Image",
                        "description": "URL of the banner image",
                        "type": "string",
                        "example": "https://example.com/banners/summer-sale.png"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the banner",
                        "type": "string",
                        "example": "This banner is for the summer sale promotion."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Banner"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Banners",
            "description": "Banners"
        }
    ]
}